+++ packages/keyring-eth-trezor/CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bump `@metamask/keyring-api` from `^23.5.0` to `^23.6.0` ([#1023](https://github.com/MetaMask/accounts/pull/1023))
-- Bump `@metamask/keyring-sdk` from `^2.2.0` to `^2.3.0` ([#1023](https://github.com/MetaMask/accounts/pull/1023))
-- Bump `@metamask/keyring-utils` from `^3.3.1` to `^4.0.0` ([#1023](https://github.com/MetaMask/accounts/pull/1023))
+- Bump `@metamask/keyring-api` from `^23.5.0` to `^23.6.0` ([#598](https://github.com/MetaMask/accounts/pull/598))
+- Bump `@metamask/keyring-sdk` from `^2.2.0` to `^2.3.0` ([#598](https://github.com/MetaMask/accounts/pull/598))
+- Bump `@metamask/keyring-utils` from `^3.3.1` to `^4.0.0` ([#598](https://github.com/MetaMask/accounts/pull/598))
 
 ## [10.1.1]